+++ src/cli/index.js
@@ -1,4 +1,3 @@
-let chalk = require('chalk')
 let fingerprint = require('@architect/utils').fingerprint
 let hydrate = require('@architect/hydrate')
 let maybeHydrate = require('../http/maybe-hydrate')
@@ -7,11 +6,11 @@ let pkgVer = require('../../package.json').version
 let ver = `Sandbox ${pkgVer}`
 let utils = require('@architect/utils')
 let watch = require('node-watch')
-let chars = utils.chars
+let updater = utils.updater
 
 // Just use Unix seperators on Windows - path.posix.normalize(process.cwd()) doesn't do what we want
 // So we normalise to slash file names (C:/foo/bar) for regex tests, etc.
-const pathToUnix = function (string) {
+let pathToUnix = function (string) {
   return string.replace(/\\/g, "/");
 }
 
@@ -44,6 +43,14 @@ module.exports = function cli(params = {}, callback) {
 
     let {arc} = utils.readArc()
     let arcFile = new RegExp(`${workingDirectory}${separator}(\\.arc|app\\.arc|arc\\.yaml|arc\\.json)`)
+    let lastEvent = Date.now()
+    let update = updater('Sandbox')
+    let deprecated = process.env.DEPRECATED
+    let ts = () => {
+      let date = new Date(lastEvent).toLocaleDateString()
+      let time = new Date(lastEvent).toLocaleTimeString()
+      console.log(`\n[${date}, ${time}]`)
+    }
 
     /**
      * Watch for pertinent filesystem changes
@@ -51,79 +58,109 @@ module.exports = function cli(params = {}, callback) {
     watcher.on('change', function (event, fileName) {
 
       // Event criteria
-      let update = event === 'update'
+      let fileUpdate = event === 'update'
       let updateOrRemove = event === 'update' || event === 'remove'
+      let ready = (Date.now() - lastEvent) >= 500
 
       fileName = pathToUnix(fileName)
 
-      let rehydrate = () => {
+      let rehydrate = ({only, msg}) => {
         let start = Date.now()
-        let status = chalk.grey('Shared file changed, rehydrating functions...')
-        console.log(`${chars.start} ${status}`)
-        hydrate.shared(() => {
+        ts()
+        update.status(msg)
+        hydrate.shared({only}, () => {
           let end = Date.now()
-          let status = chalk.grey(`Project files rehydrated into functions in ${end - start}ms`)
-          console.log(`${chars.done} ${status}`)
+          update.done(`Files rehydrated into functions in ${end - start}ms`)
         })
       }
 
       /**
        * Reload routes upon changes to Architect project manifest
        */
-      if (update && fileName.match(arcFile)) {
-        let status = chalk.grey('Architect project manifest changed, reloading HTTP routes...')
-        console.log(`${chars.start} ${status}`)
+      if (fileUpdate && fileName.match(arcFile) && ready) {
+        // TODO add arc pragma diffing, reload tables, events, etc.
+        let {arc} = utils.readArc()
 
-        let start = Date.now()
-        process.env.QUIET = true
+        // Always attempt to close the http server, but only reload if necessary
         http.close()
-        http.start(function () {
-          let end = Date.now()
-          let status = chalk.grey(`HTTP routes reloaded in ${end - start}ms`)
-          console.log(`${chars.done} ${status}`)
-        })
-        maybeHydrate(function (err) {
-          if (err) {
-            let status = chalk.grey(`Error hydrating new functions:`, err)
-            console.log(status)
-          }
-          else {
-            let status = chalk.grey(`New functions ready to go!`)
-            console.log(`${chars.done} ${status}`)
-          }
-        })
+
+        // Arc 5 only starts if it's got actual routes to load
+        let arc5 = deprecated && arc.http && arc.http.length
+        // Arc 6 may start with proxy at root, or empty `@http` pragma
+        let arc6 = !deprecated && arc.static || arc.http
+        if (arc5 || arc6) {
+          let quiet = process.env.QUIET
+          process.env.QUIET = true
+          let start = Date.now()
+          ts()
+          update.status('Architect project manifest changed, loading HTTP routes...')
+          http.start(function () {
+            let end = Date.now()
+            process.env.QUIET = quiet
+            update.done(`HTTP routes reloaded in ${end - start}ms`)
+            maybeHydrate(function (err) {
+              if (err) {
+                update.error(`Error hydrating new functions:`, err)
+              }
+              else {
+                update.done(`Functions are ready to go!`)
+                if (deprecated) {
+                  let only = 'arcFile'
+                  let msg = 'Rehydrating functions with new project manifest'
+                  rehydrate({only, msg})
+                }
+              }
+            })
+          })
+        }
+        else {
+          ts()
+          update.status('Architect project manifest changed')
+        }
       }
 
       /**
-       * Rehydrate functions with shared files upon changes to src/shared and src/views
+       * Rehydrate functions with shared files upon changes to src/shared
        */
-      if (updateOrRemove &&
-        fileName.includes(`${workingDirectory}/src/shared`) ||
-        fileName.includes(`${workingDirectory}/src/views`)) {
-        let status = chalk.grey('Shared file changed, rehydrating functions...')
-        console.log(`${chars.start} ${status}`)
-        rehydrate()
+      let isShared = fileName.includes(`${workingDirectory}/src/shared`)
+      if (updateOrRemove && ready && isShared) {
+        let only = 'shared'
+        let msg = 'Shared file changed, rehydrating functions...'
+        rehydrate({only, msg})
       }
 
+      /**
+       * Rehydrate functions with shared files upon changes to src/views
+       */
+      let isViews = fileName.includes(`${workingDirectory}/src/views`)
+      if (updateOrRemove && ready && isViews) {
+        let only = 'views'
+        let msg = 'Views file changed, rehydrating views...'
+        rehydrate({only, msg})
+      }
 
       /**
-       * Regenerate public/static.json upon changes to src/shared and src/views
+       * Regenerate public/static.json upon changes to public/
        */
-      if (updateOrRemove && arc.static &&
-          fileName.includes(`${workingDirectory}/public`) &&
-          !fileName.includes(`${workingDirectory}/public/static.json`)) {
+      let folderSetting = tuple => tuple[0] === 'folder'
+      let staticFolder = arc.static && arc.static.some(folderSetting) ? arc.static.find(folderSetting)[1] : 'public'
+      if (updateOrRemove && arc.static && ready &&
+          fileName.includes(`${workingDirectory}/${staticFolder}`) &&
+          !fileName.includes(`${workingDirectory}/${staticFolder}/static.json`)) {
         let start = Date.now()
         fingerprint({}, function next(err, result) {
-          if (err) console.log(err)
+          if (err) update.error(err)
           else {
             if (result) {
               let end = Date.now()
-              console.log(chars.done, chalk.grey(`Regenerated public/static.json in ${end - start}ms`))
-              rehydrate()
+              let only = 'staticJson'
+              let msg = `Regenerated public/static.json in ${end - start}ms`
+              rehydrate({only, msg})
             }
           }
         })
       }
+      lastEvent = Date.now()
 
     })
 
@@ -131,7 +168,7 @@ module.exports = function cli(params = {}, callback) {
      * Watch for sandbox errors
      */
     watcher.on('error', function(err) {
-      console.log(`Sandbox error:`, err)
+      update.error(`Error:`, err)
     })
   })
 }