+++ src/http/fallback.js
@@ -1,5 +1,6 @@
 let readArc = require('@architect/utils/read-arc')
-let path = require('path')
+let exists = require('fs').existsSync
+let join = require('path').join
 let {parse} = require('url')
 let invoker = require('./invoke-http')
 
@@ -13,6 +14,7 @@ module.exports = function _public(req, res, next) {
   if (req.path === '/') next()
   else {
     let arc = readArc().arc
+    let deprecated = process.env.DEPRECATED
     // reads all routes
     let routes = arc.http || []
     // add websocket route if necessary
@@ -44,25 +46,39 @@ module.exports = function _public(req, res, next) {
 
     // if either exact or wildcard match bail
     let match = exactMatch || wildMatch
-    if (match) {
+
+    // Arc v5 doesn't support implicit proxy at root, move along
+    let invalid = proxyAtRoot && deprecated
+
+    // Determine whether this is an SPA req
+    let isProxy = pathname !== '/'
+
+    if (match || invalid) {
       next()
     }
-    else if (proxyAtRoot) {
+    else if (proxyAtRoot && !deprecated) {
+      let arcProxy = join(process.cwd(), 'node_modules', '@architect', 'http-proxy', 'dist')
+      let local = join(__dirname, '..', '..', 'node_modules', '@architect', 'http-proxy', 'dist')
+      // Check to see if sandbox is being called from a local (symlink) context
+      if (exists(local)) arcProxy = local
       let exec = invoker({
         verb: 'GET',
-        pathToFunction: path.join(__dirname, '..', '..', 'vendor', 'arc-proxy-3.3.7')
+        pathToFunction: arcProxy
       })
-      req.requestContext = {}
+      if (isProxy) req.resource = '/{proxy+}'
+      else req.resource = pathname
+      req.requestContext = {} // TODO mock a {proxy+} request payload
       exec(req, res)
     }
     else {
       // invoke the get-index lambda function with a proxy payload
       let exec = invoker({
         verb: req.method.toLowerCase(),
-        pathToFunction: path.join(process.cwd(), 'src', 'http', `get-index`)
+        pathToFunction: join(process.cwd(), 'src', 'http', `get-index`)
       })
-      // TODO mock a {proxy+} request payload
-      req.requestContext = {}
+      if (isProxy) req.resource = '/{proxy+}'
+      else req.resource = pathname
+      req.requestContext = {} // TODO mock a {proxy+} request payload
       exec(req, res)
     }
   }