+++ src/invoke-lambda/runtimes/python.py
@@ -3,8 +3,9 @@ import json
 import index
 
 req = os.environ.get('__ARC_REQ__')
-context = os.environ.get('__ARC_CONTEXT__')
+con = os.environ.get('__ARC_CONTEXT__')
 event = json.loads(req)
+context = json.loads(con)
 result = '__ARC__' + json.dumps(index.handler(event, context))
 
 print(result, flush=True)