+++ src/start.js
@@ -1,4 +1,7 @@
 let chalk = require('chalk')
+let exec = require('child_process').execSync
+let exists = require('fs').existsSync
+let join = require('path').join
 let db = require('./db')
 let events = require('./events')
 let http = require('./http')
@@ -6,6 +9,7 @@ let hydrate = require('@architect/hydrate')
 let maybeHydrate = require('./http/maybe-hydrate')
 let series = require('run-series')
 let utils = require('@architect/utils')
+let updater = utils.updater
 let chars = utils.chars
 let quiet = process.env.QUIET
 
@@ -13,7 +17,9 @@ module.exports = function start(params, callback) {
   let start = Date.now()
   params = params || {}
   let {port, options, version} = params
+  let update = updater('Sandbox')
   let arc
+  let deprecated
   /**
    * Set up default sandbox port
    * CLI args > env var > passed arg
@@ -70,8 +76,7 @@ module.exports = function start(params, callback) {
     function _printBanner(callback) {
       utils.banner(params)
       if (!quiet) {
-        let msg = chalk.grey(chars.done, 'Found Architect manifest, starting up')
-        console.log(msg)
+        update.done('Found Architect manifest, starting up')
       }
       callback()
     },
@@ -80,12 +85,22 @@ module.exports = function start(params, callback) {
      * Populate additional environment variables
      */
     function _env(callback) {
-      process.env.SESSION_TABLE_NAME = 'jwe' // Default
       if (!process.env.NODE_ENV)
         process.env.NODE_ENV = 'testing'
-      if (version && version.startsWith('Architect 5'))
+      // Set Arc 5 / 6+ env
+      if (version && version.startsWith('Architect 5')) {
         process.env.DEPRECATED = true
+        deprecated = process.env.DEPRECATED
+        process.env.ARC_HTTP = 'aws'
+      }
+      else process.env.ARC_HTTP = 'aws_proxy'
+      // Read .arc-env
       utils.initEnv(callback)
+      // Populate session table (if not present)
+      if (!process.env.SESSION_TABLE_NAME)
+        process.env.SESSION_TABLE_NAME = 'jwe' // Default
+      // Declare a bucket for implicit proxy
+      process.env.ARC_STATIC_BUCKET = 'sandbox'
     },
 
     /**
@@ -101,8 +116,7 @@ module.exports = function start(params, callback) {
           if (err) callback(err)
           else {
             if (result) {
-              let msg = chalk.grey(chars.done, 'Static asset fingerpringing enabled, public/static.json generated')
-              console.log(msg)
+              update.done('Static asset fingerpringing enabled, public/static.json generated')
             }
             callback()
           }
@@ -111,10 +125,10 @@ module.exports = function start(params, callback) {
     },
 
     /**
-     *
+     * Always initialize any missing functions on startup
      */
     function _maybeInit(callback) {
-      if (!process.env.DEPRECATED) {
+      if (!deprecated) {
         utils.init(null, callback)
       }
       else callback()
@@ -130,20 +144,16 @@ module.exports = function start(params, callback) {
     /**
      * ... then hydrate Architect project files into functions
      */
-    function _maybeHydrateShared(callback) {
-      if (process.env.DEPRECATED) {
-        hydrate({install: false}, function next(err) {
-          if (err) callback(err)
-          else {
-            if (!quiet) {
-              let msg = chalk.grey(chars.done, 'Project files hydrated into functions')
-              console.log(msg)
-            }
-            callback()
+    function _hydrateShared(callback) {
+      hydrate({install: false}, function next(err) {
+        if (err) callback(err)
+        else {
+          if (!quiet) {
+            update.done('Project files hydrated into functions')
           }
-        })
-      }
-      else callback()
+          callback()
+        }
+      })
     },
 
     /**
@@ -152,8 +162,7 @@ module.exports = function start(params, callback) {
     function _db(callback) {
       client = db.start(function() {
         if (arc.tables) {
-          let msg = chalk.grey(chars.done, '@tables created in local database')
-          console.log(msg)
+          update.done('@tables created in local database')
         }
         callback()
       })
@@ -165,8 +174,7 @@ module.exports = function start(params, callback) {
     function _events(callback) {
       bus = events.start(function() {
         if (arc.events || arc.queues) {
-          let msg = chalk.grey(chars.done, '@events and @queues ready on local event bus')
-          console.log(msg)
+          update.done('@events and @queues ready on local event bus')
         }
         callback()
       })
@@ -178,8 +186,8 @@ module.exports = function start(params, callback) {
     function _http(callback) {
       let ok = () => {
         let end = Date.now()
-        let startMsg = chalk.grey(`Sandbox started in ${end - start}ms`)
-        console.log(`\n${chars.done} ${startMsg}`)
+        console.log()
+        update.done(`Sandbox started in ${end - start}ms`)
         if (!quiet) {
           let isWin = process.platform.startsWith('win')
           let ready = isWin
@@ -189,8 +197,11 @@ module.exports = function start(params, callback) {
           console.log(`${ready} ${readyMsg}`)
         }
       }
-      // two ways in
-      if (arc.static || arc.http) {
+      // Arc 5 only starts if it's got actual routes to load
+      let arc5 = deprecated && arc.http && arc.http.length
+      // Arc 6 may start with proxy at root, or empty `@http` pragma
+      let arc6 = !deprecated && arc.static || arc.http
+      if (arc5 || arc6) {
         http.start(function() {
           ok()
           let link = chalk.green.bold.underline(`http://localhost:${port}\n`)
@@ -202,6 +213,54 @@ module.exports = function start(params, callback) {
         ok()
         callback()
       }
+    },
+
+    /**
+     * Run init script (if present)
+     */
+    function _runInit(callback) {
+      let initJS = join(process.cwd(), 'scripts', 'sandbox-startup.js')
+      let initPy = join(process.cwd(), 'scripts', 'sandbox-startup.py')
+      let initRb = join(process.cwd(), 'scripts', 'sandbox-startup.rb')
+      let script
+      if (exists(initJS))
+        script = initJS
+      else if (exists(initPy))
+        script = initPy
+      else if (exists(initRb))
+        script = initRb
+      if (script) {
+        let now = Date.now()
+        update.status('Running sandbox init script')
+        let run
+        let runtime
+        if (script === initJS) {
+          // eslint-disable-next-line
+          let js = require(script)
+          run = js(arc)
+          runtime = 'Node.js'
+        }
+        else if (script === initPy) {
+          run = exec(`python ${initPy}`)
+          runtime = 'Python'
+        }
+        else {
+          run = exec(`ruby ${initRb}`)
+          runtime = 'Ruby'
+        }
+        Promise.resolve(run).then(
+          function done(result) {
+            if (result) {
+              update.done(`Init (${runtime}):`)
+              let print = result.toString().trim().split('\n').map(l => `    ${l.trim()}`).join('\n')
+              console.log(print)
+            }
+            update.done(`Sandbox init script ran in ${Date.now() - now}ms`)
+            callback()
+          }
+        )
+      }
+      else callback()
     }
   ],
   function _done(err) {
@@ -222,4 +281,3 @@ module.exports = function start(params, callback) {
 
   return promise
 }
-