+++ test/integration/http-test.js
@@ -5,14 +5,13 @@ let sandbox = require('../../src')
 let {http} = require('../../src')
 
 let b64dec = i => Buffer.from(i, 'base64').toString()
+let url = 'http://localhost:6666'
 
 let client
 test('env', t=> {
   t.plan(2)
   t.ok(sandbox, 'got sandbox')
   t.ok(http, 'got http')
-  // set the default port
-  process.env.PORT = 3333
 })
 
 /**
@@ -24,18 +23,17 @@ test('http.start', t=> {
   process.chdir(path.join(__dirname, '..', 'mock', 'normal'))
   client = http.start(function() {
     t.notOk(process.env.DEPRECATED, 'Arc v5 deprecated status NOT set')
-    t.ok(true, '@http mounted')
+    t.ok(client, '@http mounted')
   })
 })
 
 test('get /', t=> {
   t.plan(2)
-  tiny.get({
-    url: 'http://localhost:3333/'
-  }, function _got(err, data) {
+  tiny.get({url},
+  function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running nodejs10.x!'), 'is hello world')
       console.log({data})
     }
@@ -45,12 +43,12 @@ test('get /', t=> {
 test('get /binary', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/binary'
+    url: url + '/binary'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
       const img = Buffer.from(data.body).toString('base64');
-      t.ok(true, 'got /binary')
+      t.ok(data, 'got /binary')
       t.ok(img.startsWith('AAABAAEAEBAAAAEAIABoBAAAFgAAACgAAAAQAAAAIAAA'), 'is binary')
       console.log({data})
     }
@@ -60,11 +58,11 @@ test('get /binary', t=> {
 test('get /nodejs8.10', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/nodejs8.10'
+    url: url + '/nodejs8.10'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running nodejs8.10!'), 'is hello world')
       console.log({data})
     }
@@ -74,11 +72,11 @@ test('get /nodejs8.10', t=> {
 test('get /python3.7', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/python3.7'
+    url: url + '/python3.7'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running python3.7!'), 'is hello world')
       console.log({data})
     }
@@ -88,11 +86,11 @@ test('get /python3.7', t=> {
 test('get /python3.6', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/python3.6'
+    url: url + '/python3.6'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running python3.6!'), 'is hello world')
       console.log({data})
     }
@@ -102,11 +100,11 @@ test('get /python3.6', t=> {
 test('get /ruby2.5', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/ruby2.5'
+    url: url + '/ruby2.5'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running ruby2.5!'), 'is hello world')
       console.log({data})
     }
@@ -117,13 +115,13 @@ test('post /post', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.post({
-    url: 'http://localhost:3333/post',
+    url: url + '/post',
     data,
     headers: {'Content-Type': 'application/x-www-form-urlencoded'},
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'posted /post')
+      t.ok(result, 'posted /post')
       t.equal(b64dec(result.body.body), 'hi=there', 'Got base64-encoded form URL-encoded body payload')
       t.ok(result.body.isBase64Encoded, 'Got isBase64Encoded flag')
       console.log(result.body)
@@ -135,12 +133,12 @@ test('put /put', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.put({
-    url: 'http://localhost:3333/put',
+    url: url + '/put',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'put /put')
+      t.ok(result, 'put /put')
       t.equal(b64dec(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.ok(result.body.isBase64Encoded, 'Got isBase64Encoded flag')
       console.log(result.body)
@@ -156,12 +154,12 @@ test('patch /patch', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.patch({
-    url: 'http://localhost:3333/patch',
+    url: url + '/patch',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'patched /patch')
+      t.ok(result, 'patched /patch')
       t.equal(b64dec(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.ok(result.body.isBase64Encoded, 'Got isBase64Encoded flag')
       console.log(result.body)
@@ -174,12 +172,12 @@ test('delete /delete', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.del({
-    url: 'http://localhost:3333/delete',
+    url: url + '/delete',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'deleted /delete')
+      t.ok(result, 'deleted /delete')
       t.equal(b64dec(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.ok(result.body.isBase64Encoded, 'Got isBase64Encoded flag')
       console.log(result.body)
@@ -190,60 +188,68 @@ test('delete /delete', t=> {
 test('http.close', t=> {
   t.plan(1)
   client.close()
-  t.ok(true, 'http connection closed')
+  t.pass('http connection closed')
 })
 
 /**
- * Test failing to load index.html without get / defined
+ * Arc v6: test failing to load index.html without get / defined
  */
-test('http.start', t=> {
-  t.plan(2)
+let end
+test('sandbox.start', t=> {
+  t.plan(3)
   process.chdir(path.join(__dirname, '..', 'mock', 'no-index-fail'))
-  client = http.start(function() {
-    t.notOk(process.env.DEPRECATED, 'Arc v5 deprecated status NOT set')
-    t.ok(true, '@http mounted')
+  sandbox.start({}, function(err, close) {
+    if (err) t.fail(err)
+    else {
+      end = close
+      t.notOk(process.env.DEPRECATED, 'Arc v5 deprecated status NOT set')
+      t.equal(process.env.ARC_HTTP, 'aws_proxy', 'aws_proxy mode enabled')
+      t.ok(end, 'sandbox started')
+    }
   })
 })
 
 test('get / without defining get / should fail if index.html not present', t=> {
   t.plan(1)
-  tiny.get({
-    url: 'http://localhost:3333/'
-  }, function _got(err, data) {
+  tiny.get({url},
+  function _got(err, data) {
     if (err) {
-      t.equals(err.statusCode, 502, 'Got 502 for missing function')
+      t.equals(err.statusCode, 404, 'Got 404 for missing file')
     }
     else {
-      t.ok(data, 'proxy mounted')
-      console.log(data)
+      t.fail(data)
     }
   })
 })
 
-test('http.close', t=> {
+test('shut down sandbox', t=> {
   t.plan(1)
-  client.close()
-  t.ok(true, 'http connection closed')
+  end()
+  t.pass('sandbox shut down')
 })
 
 /**
- * Test successfully loading index.html without get / defined
+ * Arc v6: test successfully loading index.html without get / defined
  */
-test('http.start', t=> {
-  t.plan(2)
+test('sandbox.start', t=> {
+  t.plan(3)
   process.chdir(path.join(__dirname, '..', 'mock', 'no-index-pass'))
-  client = http.start(function() {
-    t.notOk(process.env.DEPRECATED, 'Arc v5 deprecated status NOT set')
-    t.ok(true, '@http mounted')
+  sandbox.start({}, function(err, close) {
+    if (err) t.fail(err)
+    else {
+      end = close
+      t.notOk(process.env.DEPRECATED, 'Arc v5 deprecated status NOT set')
+      t.equal(process.env.ARC_HTTP, 'aws_proxy', 'aws_proxy mode enabled')
+      t.ok(end, 'sandbox started')
+    }
   })
 })
 
 test('get / without defining get / should succeed if index.html is present', t=> {
   t.plan(1)
-  tiny.get({
-    url: 'http://localhost:3333/'
-  }, function _got(err, data) {
-    if (err) t.fail('Should not have received error')
+  tiny.get({url},
+  function _got(err, data) {
+    if (err) t.fail(err)
     else {
       t.ok(data.body.startsWith('Hello world!'), 'Proxy mounted and delivered index.html')
       console.log(data)
@@ -251,16 +257,15 @@ test('get / without defining get / should succeed if index.html is present', t=>
   })
 })
 
-test('http.close', t=> {
+test('shut down sandbox', t=> {
   t.plan(1)
-  client.close()
-  t.ok(true, 'http connection closed')
+  end()
+  t.pass('sandbox shut down')
 })
 
 /**
  * Test to ensure sandbox loads without defining @http
  */
-let end
 test('sandbox.start', t=> {
   t.plan(1)
   process.chdir(path.join(__dirname, '..', 'mock', 'no-http'))
@@ -268,16 +273,15 @@ test('sandbox.start', t=> {
     if (err) t.fail(err)
     else {
       end = close
-      t.ok(true, 'sandbox started')
+      t.ok(end, 'sandbox started')
     }
   })
 })
 
 test('get / without defining @http', t=> {
   t.plan(1)
-  tiny.get({
-    url: 'http://localhost:3333/'
-  }, function _got(err, data) {
+  tiny.get({url},
+  function _got(err, data) {
     if (err) {
       t.equals(err.code, 'ECONNREFUSED', 'Connection refused')
     }
@@ -285,10 +289,10 @@ test('get / without defining @http', t=> {
   })
 })
 
-test('sandbox.close', t=> {
+test('shut down sandbox', t=> {
   t.plan(1)
   end()
-  t.ok(true, 'http connection closed')
+  t.pass('sandbox shut down')
 })
 
 /**
@@ -302,19 +306,18 @@ test('sandbox.start', t=> {
     else {
       end = close
       t.ok(process.env.DEPRECATED, 'Arc v5 deprecated status set')
-      t.ok(true, 'sandbox started')
+      t.ok(end, 'sandbox started')
     }
   })
 })
 
 test('get /', t=> {
   t.plan(2)
-  tiny.get({
-    url: 'http://localhost:3333/'
-  }, function _got(err, data) {
+  tiny.get({url},
+  function _got(err, data) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'got /')
+      t.ok(data, 'got /')
       t.ok(data.body.startsWith('Hello from Architect Sandbox running nodejs10.x!'), 'is hello world')
       console.log({data})
     }
@@ -324,12 +327,12 @@ test('get /', t=> {
 test('get /binary', t=> {
   t.plan(2)
   tiny.get({
-    url: 'http://localhost:3333/binary'
+    url: url + '/binary'
   }, function _got(err, data) {
     if (err) t.fail(err)
     else {
       const img = Buffer.from(data.body).toString('base64');
-      t.ok(true, 'got /binary')
+      t.ok(data, 'got /binary')
       t.ok(img.startsWith('AAABAAEAEBAAAAEAIABoBAAAFgAAACgAAAAQAAAAIAAA'), 'is binary')
       console.log({data})
     }
@@ -340,17 +343,16 @@ test('post /post', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.post({
-    url: 'http://localhost:3333/post',
+    url: url + '/post',
     data,
     headers: {'Content-Type': 'application/x-www-form-urlencoded'},
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'posted /post')
+      t.ok(result, 'posted /post')
       t.equal(JSON.stringify(result.body.body), JSON.stringify(data), 'Got base64-encoded form URL-encoded body payload')
       t.notOk(result.body.isBase64Encoded, 'No isBase64Encoded flag')
       console.log(result.body)
-      t.end()
     }
   })
 })
@@ -359,16 +361,15 @@ test('put /put', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.put({
-    url: 'http://localhost:3333/put',
+    url: url + '/put',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'put /put')
+      t.ok(result, 'put /put')
       t.equal(JSON.stringify(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.notOk(result.body.isBase64Encoded, 'No isBase64Encoded flag')
       console.log(result.body)
-      t.end()
     }
   })
 })
@@ -381,16 +382,15 @@ test('patch /patch', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.patch({
-    url: 'http://localhost:3333/patch',
+    url: url + '/patch',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'patched /patch')
+      t.ok(result, 'patched /patch')
       t.equal(JSON.stringify(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.notOk(result.body.isBase64Encoded, 'No isBase64Encoded flag')
       console.log(result.body)
-      t.end()
     }
   })
 })
@@ -400,22 +400,21 @@ test('delete /delete', t=> {
   t.plan(3)
   let data = {hi: 'there'}
   tiny.del({
-    url: 'http://localhost:3333/delete',
+    url: url + '/delete',
     data,
   }, function _got(err, result) {
     if (err) t.fail(err)
     else {
-      t.ok(true, 'deleted /delete')
+      t.ok(result, 'deleted /delete')
       t.equal(JSON.stringify(result.body.body), JSON.stringify(data), 'Got base64-encoded JSON-encoded body payload')
       t.notOk(result.body.isBase64Encoded, 'No isBase64Encoded flag')
       console.log(result.body)
-      t.end()
     }
   })
 })
 
-test('sandbox.close', t=> {
+test('shut down sandbox', t=> {
   t.plan(1)
   end()
-  t.ok(true, 'http connection closed')
+  t.pass('sandbox shut down')
 })