+++ changelog.md
@@ -2,6 +2,76 @@
 
 ---
 
+## [1.4.0] 2019-09-26
+
+### Added
+
+- Auto-hydration received a bunch of nice upgrades:
+  - Auto-hydration now detects changes to the state of your installed Node dependencies, and rehydrates if necessary; for example:
+    - You're working on a project, and a teammate updates a dependency in `get /foo` from version `1.0.0` to `1.1.0`
+    - Upon your next git pull, `sandbox` will detect the dependency update in `get /foo` and automatically install version `1.1.0` for you
+  - Auto-hydration now has a rate limit of one change every 500ms to prevent recursive or aggressive file updates
+  - Auto-hydration now has `@static folder` support
+  - Auto-hydration now only hydrates the shared files necessary
+    - For example: if you change a file in `src/views`, it will only update your `@views` functions, and not attempt to rehydrate all your project's functions with `src/shared`
+  - Events now have a timestamp and improved formatting
+- Beta: `sandbox` init script support!
+  - `sandbox` will now run the init script of your choosing upon startup after all subsystems have started up:
+    - `scripts/sandbox-startup.js` - a CommonJS module, receives your parsed Arc project as a parameter, supports async/await
+    - `scripts/sandbox-startup.py` - a Python script
+    - `scripts/sandbox-startup.rb` - a Ruby script
+
+
+### Changed
+
+- Improvements to auto-hydration of src/shared and src/views upon startup
+- Improvements to the conditions under which the HTTP server starts, shuts down, and restarts
+- Improved async error copy (displayed when execution does not complete)
+- Proxied requests now sends a proper `req.resource`, which can resolve some SPA bugs, esp when used with newer Arc Functions
+- `sandbox` now respects and errors on invalid response params for proper Architect 6 compatibility, fixes #49
+- Updates Dynalite to `3.0.0`, thanks @mhart!
+- Better 404 / file missing handling in `sandbox` when using `http.proxy` (or loading assets without `@http get /` specified)
+
+
+### Fixed
+
+- Fixed issue where in certain circumstances `get /` wouldn't reload after a change to the project manifest
+- Minor fix where if you specified a `SESSION_TABLE_NAME` env var outside of `.arc-env`, `sandbox` won't clobber it
+- Fixed caching headers for various error states (async, timeout, etc.) to ensure your browser won't accidentally cache an error response
+
+---
+
+## [1.3.14] 2019-09-16
+
+### Changed
+
+- Internal change: moves vendored HTTP proxy bundle into node module
+- This change incorporates a number of fixes by way of `@architect/functions` (see below)
+
+
+### Fixes
+
+- Fixes issue where binary assets delivered via `sandbox` / root may not be properly encoded
+- Fixes issue where `http.proxy.public` + `http.proxy.read` may not have delivered correctly formatted responses in an Architect 5 environment
+
+---
+
+## [1.3.13] 2019-09-15
+
+### Fixed
+
+- Fixes callback error when auto-rehydrating `src/shared` and `src/views`
+
+---
+
+## [1.3.12] 2019-09-13
+
+### Fixed
+
+- Fixes empty mock `context` object encoding
+
+---
+
 ## [1.3.10 - 1.3.11] 2019-09-09
 
 ### Fixed